Excercises till 4.22

+++ osa4/blogilista/controllers/blogs.js
@@ -1,13 +1,17 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
+  console.log(console.log(`token's id ${request.token.id}`))
+  console.log(console.log(`token's user ${request.user}`))
   if (body.title === undefined || body.url === undefined) {
     return response.status(400).end()
   } 
@@ -16,16 +20,25 @@ blogsRouter.post('/', async (request, response) => {
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes || 0
+      likes: body.likes || 0,
+      user: user._id,
     })
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     return response.status(201).json(savedBlog)  
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  return response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  } 
+  return response.status(401).json({ error: 'not allowed. user is not the owner' })
 })
 
 blogsRouter.get('/:id', async (request, response) => {
